Link `ftPr_SpecialHi` (1.22 KB) (#1327)
