+++ src/melee/ft/ftdata.c
@@ -177,6 +177,7 @@
 #include "ftPopo/ftPp_SpecialS.h"
 #include "ftPurin/ftPr_Init.h"
 #include "ftPurin/ftPr_SpecialHi.h"
+#include "ftPurin/ftPr_SpecialLw.h"
 #include "ftSamus/ftSs_Init.h"
 #include "ftSamus/ftSs_SpecialHi.h"
 #include "ftSamus/ftSs_SpecialLw_1.h"
